Connect every row of the DFAs to the accepting state.
The two locations where DFAs are constructed for representing a collection of
TCRs -- in positive selection against a peptide and in constructing the whole
possible repertoire -- both contain tests for whether edges to the accepting
state should be added. These tests are conditioned on the variable l, which is the
current row number being constructed, being equal to r or r-1, r being the
total number of rows. As a result, transitions to the accepting state are only
added in the final two rows, such that the behavior of the resulting DFA is to
accept only when there are /exactly/ r mismatches.

The proposed fix is to change the equality tests to inequality tests.
Transitions are added from the end of every row to the accepting state and the
resulting DFA accepts matches with r or less mismatches.